update employee edit

+++ src/app/components/employee/edit-employee/edit-employee.component.ts
@@ -3,8 +3,6 @@ import {
   Input,
   OnChanges,
   OnDestroy,
-  OnInit,
-  Output,
   SimpleChanges,
   TemplateRef,
   ViewChild
@@ -13,12 +11,12 @@ import { FormsModule } from '@angular/forms';
 import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { EmployeeDataModalComponent } from "../../modal/employee-data-modal/employee-data-modal.component";
 import { EditEmployeeService } from '../../services/EmployeeEditService';
-import {AsyncPipe, NgForOf, NgIf} from "@angular/common";
-import {DataService} from "../../../service/data.service";
-import {Observable, Subscription} from "rxjs";
-import {Employee} from "../../../model/Employee";
-import {AddQualificationService} from "../../services/AddQualificationService";
-import {AddEmployee} from "../../../model/AddEmployee";
+import { AsyncPipe, NgForOf, NgIf } from "@angular/common";
+import { DataService } from "../../../service/data.service";
+import { Observable, Subscription } from "rxjs";
+import { Employee } from "../../../model/Employee";
+import { AddQualificationService } from "../../services/AddQualificationService";
+import { AddEmployee } from "../../../model/AddEmployee";
 
 @Component({
   selector: 'app-edit-employee',
@@ -27,70 +25,20 @@ import {AddEmployee} from "../../../model/AddEmployee";
   templateUrl: './edit-employee.component.html',
   styleUrls: ['./edit-employee.component.css']
 })
-export class EditEmployeeComponent  implements OnChanges, OnDestroy{
-
+export class EditEmployeeComponent implements OnChanges, OnDestroy {
   @ViewChild(EmployeeDataModalComponent) modal!: EmployeeDataModalComponent;
-  title: string = "Mitarbeiter bearbeiten";
-  qualificationToRemove: number | null = null;
-  @ViewChild('deleteQualificationModal', {static: true}) deleteQualificationModal!: TemplateRef<any>;
-  protected modalRef!: NgbModalRef;
-  employee$!: Observable<Employee>;
+  @ViewChild('deleteQualificationModal', { static: true }) deleteQualificationModal!: TemplateRef<any>;
+
   @Input() employeeId?: number;
-  employee: Employee | null = null; // Lokale Kopie
 
-  private subscriptions: Subscription = new Subscription();
-  private qid: number | undefined;
-  private eid: number | undefined;
-
-  constructor(protected editEmployeeService: EditEmployeeService, private modalService: NgbModal, private dataService: DataService,  private addQualificationService: AddQualificationService) {
-  }
-
-  ngOnChanges(changes: SimpleChanges): void {
-    if (changes['employeeId'] && changes['employeeId'].currentValue) {
-      this.loadEmployee();
-    }
-  }
-
-  private loadEmployee(): void {
-    if (this.employeeId !== undefined) {
-      this.employee$ = this.dataService.getEmployee(this.employeeId);
-      this.employee$.subscribe((employee: Employee) => {
-        this.employee = { ...employee };
-      });
-    }
-  }
-
-  onSaveChanges() {
-    this.employee$.subscribe((employee: Employee) => {
-
-      // @ts-ignore
-      this.employeeAdd.id = this.employee.id
-      // @ts-ignore
-      this.employeeAdd.firstName = this.employee.firstName;
-      // @ts-ignore
-      this.employeeAdd.lastName = this.employee.lastName;
-      // @ts-ignore
-      this.employeeAdd.city = this.employee.city;
-      // @ts-ignore
-      this.employeeAdd.phone = this.employee.phone;
-      // @ts-ignore
-      this.employeeAdd.postcode = this.employee.postcode;
-      // @ts-ignore
-      this.employeeAdd.street = this.employee.street;
-      // @ts-ignore
-      this.employeeAdd.skillSet = this.employee.skillSet.map(skill =>
-        typeof skill === "object" && skill.id !== undefined ? skill.id : skill
-      ).filter(id => typeof id === "number");
-
-      console.log(this.employee);
-
-      this.dataService.updateEmployee(this.employeeAdd).subscribe(response => {
-        console.log('Update erfolgreich:', response);
-      });
-    });
-    this.modal.closeModal();
-    this.editEmployeeService.setValue(false);
-  }
+  title: string = "Mitarbeiter bearbeiten";
+  modalRef!: NgbModalRef;
+  employee$!: Observable<Employee>;
+  employee: Employee | null = null;
+  qualificationToRemove: number | null = null;
+  private subscriptions = new Subscription();
+  private eid?: number;
+  private qid?: number;
 
   employeeAdd: AddEmployee = {
     lastName: "",
@@ -102,37 +50,95 @@ export class EditEmployeeComponent  implements OnChanges, OnDestroy{
     skillSet: []
   };
 
+  constructor(
+    protected editEmployeeService: EditEmployeeService,
+    private modalService: NgbModal,
+    private dataService: DataService,
+    private addQualificationService: AddQualificationService
+  ) {}
+
+  ngOnChanges(changes: SimpleChanges): void {
+    if (changes['employeeId']?.currentValue) {
+      this.loadEmployee();
+    }
+  }
+
+  private loadEmployee(): void {
+    if (this.employeeId !== undefined) {
+      this.employee$ = this.dataService.getEmployee(this.employeeId);
+      this.subscriptions.add(
+        this.employee$.subscribe(employee => {
+          this.employee = { ...employee };
+        })
+      );
+    }
+  }
+
+  onSaveChanges(): void {
+    if (!this.employee) return;
+
+    this.employeeAdd = {
+      ...this.employeeAdd,
+      id: this.employee.id,
+      firstName: this.employee.firstName,
+      lastName: this.employee.lastName,
+      city: this.employee.city,
+      phone: this.employee.phone,
+      postcode: this.employee.postcode,
+      street: this.employee.street,
+      skillSet: [
+        // @ts-ignore
+        ...this.employeeAdd.skillSet,
+        // @ts-ignore
+        ...this.employee.skillSet
+          .map(skill => (typeof skill === "object" && skill.id !== undefined ? skill.id : skill))
+          .filter((id): id is number => typeof id === "number")
+          // @ts-ignore
+          .filter(id => !this.employeeAdd.skillSet.includes(id))
+      ]
+    };
+
+    this.dataService.updateEmployee(this.employeeAdd).subscribe(() => {
+      console.log('Update erfolgreich');
+    });
+
+    this.closeModal();
+  }
+
   ngOnDestroy(): void {
     this.subscriptions.unsubscribe();
   }
 
-  closeModal() {
+  closeModal(): void {
     this.modal.closeModal();
     this.editEmployeeService.setValue(false);
   }
 
-  onAddQualificationClick (){
-    this.addQualificationService.setValue(true)
-    this.addQualificationService.setEmployee(this.employeeAdd)
+  onAddQualificationClick(): void {
+    this.addQualificationService.setValue(true);
+    this.addQualificationService.setEmployee(this.employeeAdd);
   }
 
-  setQualificationToRemove(qid: number | undefined, eid: number | undefined) {
-    this.eid = eid;
+  setQualificationToRemove(qid?: number, eid?: number): void {
     this.qid = qid;
+    this.eid = eid;
     this.openDeleteModal();
   }
 
-  openDeleteModal() {
-    this.modalRef = this.modalService.open(this.deleteQualificationModal, {ariaLabelledBy: 'deleteModalLabel'});
+  openDeleteModal(): void {
+    this.modalRef = this.modalService.open(this.deleteQualificationModal, { ariaLabelledBy: 'deleteModalLabel' });
   }
 
-  confirmDelete() {
+  confirmDelete(): void {
+    if (this.eid === undefined || this.qid === undefined) return;
+
     this.dataService.deleteQualificationFromEmployee(this.eid, this.qid).subscribe({
       next: () => {
         this.modalRef.close();
-      },
-      error: (err) => console.error('Fehler beim Löschen:', err),
+        this.loadEmployee();
+        },
+      error: err => console.error('Fehler beim Löschen:', err)
     });
-  }
 
+  }
 }
